+++ apps/server/test/utils/index.ts
@@ -1,14 +1,1 @@
-import axios from 'axios'
-
-export class postman {
-  static backend = 'http://localhost:8080'
-  static url = postman.backend + '/api'
-
-  static async root() {
-    return axios.get(postman.backend)
-  }
-
-  static async get(endpoint: string, params?: any) {
-    return axios.get(postman.url + endpoint, { params })
-  }
-}
+export * from './postman'
